- Added first version of a coordinator that gets triggered when a mapper puts its output in a defined bucket - The coordinator detects when a map phase is done and puts a marker in the status bucket

+++ mapreduce/src/main/java/driver/MappersInfo.java
@@ -3,30 +3,30 @@ package driver;
 import java.util.Map;
 
 public class MappersInfo {
-    private long mapperCount;
-    private Map<Long, Integer> batchCountPerMapper;
+    private int mapperCount;
+    private Map<Integer, Integer> batchCountPerMapper;
 
     public MappersInfo() {
     }
 
-    public MappersInfo(long mapperCount, Map<Long, Integer> batchCountPerMapper) {
+    public MappersInfo(int mapperCount, Map<Integer, Integer> batchCountPerMapper) {
         this.mapperCount = mapperCount;
         this.batchCountPerMapper = batchCountPerMapper;
     }
 
-    public long getMapperCount() {
+    public int getMapperCount() {
         return mapperCount;
     }
 
-    public void setMapperCount(long mapperCount) {
+    public void setMapperCount(int mapperCount) {
         this.mapperCount = mapperCount;
     }
 
-    public Map<Long, Integer> getBatchCountPerMapper() {
+    public Map<Integer, Integer> getBatchCountPerMapper() {
         return batchCountPerMapper;
     }
 
-    public void setBatchCountPerMapper(Map<Long, Integer> batchCountPerMapper) {
+    public void setBatchCountPerMapper(Map<Integer, Integer> batchCountPerMapper) {
         this.batchCountPerMapper = batchCountPerMapper;
     }
 }
